+++ lib/crowdin_sdk.dart
@@ -1,5 +1,4 @@
-library crowdin_sdk;
-
 export 'src/crowdin.dart';
 export 'src/real_time_preview/crowdin_auth_config.dart';
+export 'src/real_time_preview/crowdin_preview_state.dart';
 export 'src/real_time_preview/crowdin_real_time_preview_widget.dart';